+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 
 * ![Bugfix][badge-bugfix] Documenter now generates the correct source URLs for docstrings from other packages when the `repo` argument to `makedocs` is set (note: the source links to such docstrings only work if the external package is cloned from GitHub and added as a dev-dependency). However, this change **breaks** the case where the `repo` argument is used to override the main package/repository URL, assuming the repository is cloned from GitHub. ([#1808][github-1808])
 
+* ![Enhancement][badge-enhancement] Added Woodpecker CI as another CI solution for deploying docs. ([#1880][github-1880])
+
 ## Version `v0.27.21`
 
 * ![Bugfix][badge-bugfix] Fix a regression where Documenter throws an error on systems that do not have Git available. ([#1870][github-1870], [#1871][github-1871])
@@ -1102,6 +1104,7 @@
 [github-1865]: https://github.com/JuliaDocs/Documenter.jl/pull/1865
 [github-1870]: https://github.com/JuliaDocs/Documenter.jl/issues/1870
 [github-1871]: https://github.com/JuliaDocs/Documenter.jl/pull/1871
+[github-1880]: https://github.com/JuliaDocs/Documenter.jl/pull/1880
 <!-- end of issue link definitions -->
 
 [julia-38079]: https://github.com/JuliaLang/julia/issues/38079